wgpu: Moved WgpuRenderBackend into backend.rs
