+++ exporter/src/main.rs
@@ -5,10 +5,11 @@ use indicatif::{ProgressBar, ProgressStyle};
 use rayon::prelude::*;
 use ruffle_core::tag_utils::SwfMovie;
 use ruffle_core::PlayerBuilder;
+use ruffle_render_wgpu::backend::WgpuRenderBackend;
 use ruffle_render_wgpu::clap::{GraphicsBackend, PowerPreference};
 use ruffle_render_wgpu::descriptors::Descriptors;
 use ruffle_render_wgpu::target::TextureTarget;
-use ruffle_render_wgpu::{wgpu, WgpuRenderBackend};
+use ruffle_render_wgpu::wgpu;
 use std::fs::create_dir_all;
 use std::panic::catch_unwind;
 use std::path::{Path, PathBuf};
